many fixes to make jdr.sh work
fixed several little issues with paths in the jdr.sh script
factored out AS7 plugin code to promote re-use for other products
